split into multiple components
