+++ apps/minaintyg/src/pages/Certificates/components/CertificateCard/CertificateCardEvents/CertificateCardEventsInfo.tsx
@@ -1,0 +1,18 @@
+import { useFormat } from '../../../../../hooks/useFormat'
+import { CertificateListEvent } from '../../../../../schema/certificateList.schema'
+
+export function CertificateCardEventsInfo({ events }: { events: CertificateListEvent[] }) {
+  const { formatDatetime } = useFormat()
+  return (
+    <div>
+      <div className="font-bold">
+        Senaste händelser (visar {events.length} av {events.length} händelser)
+      </div>
+      {events.map(({ timestamp, description }) => (
+        <div key={timestamp}>
+          {formatDatetime(timestamp)} {description}
+        </div>
+      ))}
+    </div>
+  )
+}
